src/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
django-environ==0.4.4
-e git+https://github.com/mobolic/facebook-sdk.git@27697ed3743e3a60da6907267280755d324a85f3#egg=facebook_sdk
idna==2.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
dj-database-url==0.4.2
gunicorn==19.7.1
postgres==2.2.1
whitenoise==3.3.0
phonenumbers
django-rest-framework==0.1.0
djangorestframework==3.8.2
django-nose
pyotp==2.2.6
